Move DiagnosticsExtensions to shared generation project
